:recycle: refactor: repository 폴더로 GroupCategoryRepositoy 이동
