+++ backend/src/app/group-article/group-article.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { GroupCategoryRepository } from '@app/group-article/group-category.repository';
+import { GroupCategoryRepository } from '@app/group-article/repository/group-category.repository';
 import { GroupArticleController } from '@app/group-article/group-article.controller';
 
 @Module({
