+++ backend/src/app/group-article/repository/group-category.repository.ts
